Moved columns to data source, lint errors
