+++ custom/ilsgateway/zipline/data_sources/zipline_warehouse_order_data_source.py
@@ -1,3 +1,6 @@
+from django.utils.translation import ugettext as _
+
+from corehq.apps.reports.datatables import DataTablesColumn
 from custom.ilsgateway.zipline.data_sources.zipline_data_source import ZiplineDataSource
 from custom.zipline.models import EmergencyOrder
 from custom.ilsgateway.zipline import helpers
@@ -38,6 +41,41 @@ class ZiplineWarehouseOrderDataSource(ZiplineDataSource):
     @property
     def total_count(self):
         return EmergencyOrder.objects.filter(**self.filters).count()
+
+    @property
+    def columns(self):
+        return [
+            DataTablesColumn('order id', help_text=_('unique identifier for each order assigned by ILSGateway')),
+            DataTablesColumn('phone number', help_text=_('phone number that submitted the emg request')),
+            DataTablesColumn('date', help_text=_('timestamp for receipt of incoming emg request, automatic')),
+            DataTablesColumn('location code', help_text=_('the location that corresponds to the health facility')),
+            DataTablesColumn('status', help_text=_('current status of the transaction (rejected, cancelled, '
+                                                   'cancelled by user, received, approved, dispatched, delivered, '
+                                                   'confirmed)')),
+            DataTablesColumn('delivery lead time', help_text=_('time between Request to Zipline and delivery '
+                                                               'completed, in minutes (Time for zipline '
+                                                               'to get request and then to deliver)')),
+            DataTablesColumn('Request Attempts', help_text=_('the number of times ILS gateway '
+                                                             'tried to submit a request to the zipline API')),
+            DataTablesColumn('status.received', help_text=_('timestamp of received status '
+                                                            '(request forwarded to zipline)')),
+            DataTablesColumn('status.rejected', help_text=_('timestamp of rejected status '
+                                                            '(request forwarded to zipline and zipline rejects)')),
+            DataTablesColumn('status.cancelled', help_text=_('timestamp from Zipine status update '
+                                                             'if request is closed (cancelled) at level 3. ')),
+            DataTablesColumn('status.approved', help_text=_('timestamp for approval of emg '
+                                                            'request by Zipline warehouse, automated')),
+            DataTablesColumn('status.dispatched', help_text=_('timestamp of dispatch of first vehicle')),
+            DataTablesColumn('status.delivered', help_text=_('time stamp for delivery of final vehicle')),
+            DataTablesColumn('products requested', help_text=_('structured string with product long codes '
+                                                               '(for example, 10010203MD) and quantities '
+                                                               'for products requested in emg request ')),
+            DataTablesColumn('products delivered', help_text=_('structured string with product '
+                                                               'codes for products delivered by zipline')),
+            DataTablesColumn('products requested and not confirmed',
+                             help_text=_('structured string with products '
+                                         'that were not confirmed based on the request'))
+        ]
 
     def get_data(self, start, limit):
         emergency_orders = self.get_emergency_orders(start, limit)
