+++ custom/ilsgateway/zipline/reports/zipline_warehouse_package_report.py
@@ -1,8 +1,5 @@
 from collections import namedtuple
 
-from django.utils.translation import ugettext as _
-
-from corehq.apps.reports.datatables import DataTablesHeader, DataTablesColumn
 from corehq.apps.reports.filters.dates import DatespanFilter
 from corehq.apps.reports.filters.fixtures import AsyncLocationFilter
 from custom.ilsgateway.zipline.data_sources.zipline_warehouse_package_data_source import \
@@ -41,27 +38,6 @@ class ZiplineWarehousePackageReport(ZiplineReport):
             location_id=self.location_id,
             statuses=self.statuses,
             orders_id=self.orders_id
-        )
-
-    @property
-    def headers(self):
-        return DataTablesHeader(
-            DataTablesColumn('Order id', help_text=_('unique id assigned to the order by ILSGateway')),
-            DataTablesColumn('Location code', help_text=_('the location that corresponds to the health facility')),
-            DataTablesColumn('Status',
-                             help_text=_('"current status of the transaction (dispatched cancelled delivered)"')),
-            DataTablesColumn('Status dispatched', help_text=_('time that uav is launched to delivery site')),
-            DataTablesColumn('Status delivered', help_text=_('time that vehicle dropped package')),
-            DataTablesColumn('Delivery leadtime', help_text=_('difference between dispatched and delivered')),
-            DataTablesColumn('Package number', help_text=_('a sequential number assigned '
-                                                           'to each package within an order')),
-            DataTablesColumn('Vehicle id', help_text=_('the unique id for the vehicle that is set to be delivered,'
-                                                       ' will be repeated based on vehciles at warehouse')),
-            DataTablesColumn('Package id', help_text=_('the unique id for the package '
-                                                       'that is set to be delivered')),
-            DataTablesColumn('Package weight (grams)', help_text=_('calculated weight of'
-                                                                   ' the products in the vehicle')),
-            DataTablesColumn('Products in package')
         )
 
     @property
